S12P21A408-master/AI/chatbot/requirements.txt
Flask
requests
beautifulsoup4
aiohttp
yake
flask-cors
pymysql
fastapi
uvicorn
pydantic
python-dotenv
dbutils
mysql-connector-python
pyjwt
tldextract
keybert
sentence-transformers==4.0.1
transformers==4.50.3
torch
konlpy
mecab-python3==1.0.10
chardet